Move AuthenticatorBridgeConnectionType extension to its own package and file
