+++ authenticatorbridge/src/main/java/com/bitwarden/authenticatorbridge/manager/model/AuthenticatorBridgeConnectionType.kt
@@ -17,12 +17,3 @@ enum class AuthenticatorBridgeConnectionType {
      */
     DEV,
 }
-
-/**
- * Convert a [AuthenticatorBridgeConnectionType] to raw package name for connection.
- */
-internal fun AuthenticatorBridgeConnectionType.toPackageName() =
-    when (this) {
-        AuthenticatorBridgeConnectionType.RELEASE -> "com.x8bit.bitwarden"
-        AuthenticatorBridgeConnectionType.DEV -> "com.x8bit.bitwarden.dev"
-    }
